feat(buckets): adds simplified open method without thread requirements
