+++ packages/buckets/src/index.ts
@@ -1,1 +1,2 @@
 export * from './buckets'
+export { InitReply, Root } from '@textile/buckets-grpc/buckets_pb'
