+++ packages/hub-threads-client/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@textile/buckets",
-  "version": "0.2.6",
-  "description": "web-gRPC wrapper client for managing Textile Buckets.",
+  "name": "@textile/hub-threads-client",
+  "version": "0.0.1",
+  "description": "web-gRPC wrapper client for communicating with Threads APIs on the Hub.",
   "main": "dist/index",
   "types": "dist/index",
   "publishConfig": {
@@ -29,20 +29,18 @@
   "license": "MIT",
   "dependencies": {
     "@improbable-eng/grpc-web": "^0.12.0",
-    "@textile/buckets-grpc": "^1.0.8",
     "@textile/context": "^0.4.8",
-    "@textile/hub-threads-client": "^0.0.1",
     "@textile/security": "^0.1.5",
     "@textile/threads-client": "^0.6.14",
+    "@textile/threads-id": "^0.1.7",
+    "@textile/users-grpc": "^1.0.8",
     "cids": "^0.8.0",
-    "event-iterator": "^2.0.0",
-    "loglevel": "^1.6.8"
+    "loglevel": "^1.6.8",
+    "multihashes": "0.4.19"
   },
   "devDependencies": {
     "@textile/hub-grpc": "^1.0.8",
-    "@textile/threads-client": "^0.6.14",
     "@textile/threads-core": "^0.1.25",
-    "@textile/threads-id": "^0.1.7",
     "axios": "^0.19.2",
     "delay": "^4.3.0",
     "fast-sha256": "^1.3.0",
