+++ packages/hub-threads-client/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * A web-gRPC wrapper client for communicating with the web-gRPC enabled Threads ove the Hub APIs.
+ * @packageDocumentation
+ */
+
+export { Client } from './client'
