--- packages/hub/lib/hub.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = hub;
-
-function hub() {
-    // TODO
-}
