+++ packages/hub/src/index.ts
@@ -7,4 +7,4 @@ export { ThreadID, Variant } from '@textile/threads-id'
 export { Identity } from '@textile/threads-core'
 export * from '@textile/buckets'
 export * from '@textile/security'
-export { Client } from './users'
+export { Client } from '@textile/hub-threads-client'
